requirements.txt
Flask
pandas
matplotlib
scikit-learn
reportlab
psycopg2-binary
